MAINT: move conj_exponent to utility and remove some obsolete imports
